+++ cases/other_profile_cases.py
@@ -38,6 +38,7 @@ class OtherProfileTest(Test):
 
         self.login()
         self.page.open_other_profile(self.ID2)
+        self.page.refresh()
         subscribe_text = self.page.get_subscribe_button_text()
         self.page.click_subscribe()
         unsubscribe_text = self.page.get_subscribe_button_text()
